backend/requirements.txt
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
psycopg2-binary==2.9.1

